Create Saving Lives
The input consists of 
 different segments:

1. The first line will have the number of test cases t, where 1 <= t <= 10.

2. The first line of each case will contain the first positive numerically distanced whole number, which will have the number’s digits, 0–9, be space-separated. The amount of digits in this line is m, where 1 <= m <= 9.

The second line of each case will contain the second positive numerically distanced whole number, which will have the number’s digits, 0 – 9, be space-separated. The amount of digits in this line is m, where 1 <= n <= 9.

Output
Output the sum of the numerically distanced numbers for each of the test cases on one line each.